+++ assets/js/octantis/index.js
@@ -53,4 +53,13 @@ const ShopifyToastHook = {
   }
 }
 
-export { OctantisInteractable, ShopifyAppBridgeModal, ShopifyModal, ShopifyToastHook }
+export const AppBridgeNavManu = {
+  mounted(){
+    this.el.addEventListener('click', (event) => {
+      event.preventDefault();
+      this.liveSocket.execJS(this.el, this.el.getAttribute("data-nav-event"));
+    });
+  }
+}
+
+export { AppBridgeNavManu, OctantisInteractable, ShopifyAppBridgeModal, ShopifyModal, ShopifyToastHook }